Criar services e controllers

+++ src/main/java/com/meubolso/repositories/TipoCombustivelRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TipoCombustivelRepository extends JpaRepository<TipoCombustivel, Integer> {
+public interface TipoCombustivelRepository extends JpaRepository<TipoCombustivel, Long> {
 }
